+++ docs/CHANGELOG.md
@@ -18,6 +18,8 @@ Changes are grouped as follows
 - Support for deleting by `NodeId` in generated SDK.
 - You can now use `DirectRelationReference` and `tuple[str, str]` as node identifiers in the Write classes
   of the generated SDK.
+- In the `.list` method, `NodeId` and `DirectRelationReference` can now be used when filtering on direct relations
+  (in addition to `tuple[str, str]` and `str`).
 
 ## [0.99.36] - 24-10-04
 ### Changed